\name{causalcp-package}
\alias{causalcp-package}
\docType{package}
\title{Causal CP-Plot Diagnostics}
\description{
Tools for drawing CATE-propensity-score plots and local IV
complier-weighted CP plots as diagnostic visualizations. The package
accompanies Tian, Yang, and Ding (2026), "Bracketing Relationships of Weighted
Average Treatment Effects", arXiv:2606.11715.
}
