% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{cc_loss}
\alias{cc_loss}
\title{Causal change loss function.}
\usage{
cc_loss(
  X,
  Y,
  i,
  st,
  en,
  min_seg = 20,
  cond_sets = get_powerset(colnames(X)),
  intercept = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{X}{The covariate matrix of dimension \verb{n x d}.}

\item{Y}{The response vector of length \code{n}.}

\item{i}{The time index at which to evaluate the loss, relative to \code{st}.}

\item{st}{The start index (inclusive).}

\item{en}{The end index (inclusive).}

\item{cond_sets}{The conditioning sets. Default to the power set of all covariates.}

\item{intercept}{Whether to include an intercept in the linear regressions.}

\item{split_setting}{A list of two elements: \code{type} and \code{n_split}. \code{type} can
be one of "one_vs_rest" or "rest_vs_one"; \code{n_split} is the number of splits
over the interval \code{st:(st+i-1)} and \code{(st+i):n} for calculating the loss.}
}
\value{
A scalar value.
}
\description{
Causal change loss function.
}
\examples{
tmp <- sapply(eval_grid, \(t) cc_loss(X, Y, t, 1, n,
  split_setting = list(type = "one_vs_rest", n_split  = 3), cond_sets = PS))

}
