% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{cc_loss_one_side}
\alias{cc_loss_one_side}
\title{Calculate the causal change loss on one side of the split.}
\usage{
cc_loss_one_side(
  X,
  Y,
  split_lst,
  cond_sets = get_powerset(colnames(X)),
  intercept = TRUE
)
}
\arguments{
\item{split_lst}{A list of splits. Each element is a list of two vectors of
indices, where the 1st set of indices can be seen as the training set
and the 2nd set can be seen as the test set.}

\item{cond_sets}{The conditioning sets. Default to the power set of all covariates.}

\item{intercept}{Whether to include an intercept in the linear regressions.}
}
\value{
A scalar value.
}
\description{
Calculate the causal change loss on one side of the split.
}
