% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{comp_cond_dist}
\alias{comp_cond_dist}
\title{Compute the conditional distribution (mean and covariance) based on graph G.}
\usage{
comp_cond_dist(mu, Sig, dep_idx, cond_idx, cond_val)
}
\arguments{
\item{mu}{A vector of the joint mean.}

\item{Sig}{A matrix of the joint covariance.}

\item{dep_idx}{Indices of the dependent variables.}

\item{cond_idx}{Indices of the conditioning variables.}

\item{cond_val}{Values of the conditioning variables. It should have the
same length as \code{cond_idx}.}
}
\value{
A list of two elements: the conditional mean \code{mu} and the conditional
covariance matrix \code{Sig}.
}
\description{
Compute the conditional distribution (mean and covariance) based on graph G.
}
