% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{comp_joint_cov}
\alias{comp_joint_cov}
\title{Compute the theoretical covariance matrix based of graph G.}
\usage{
comp_joint_cov(a12, b15, b25, a53, a43, s1, s2, s3, s4, s5)
}
\description{
Compute the theoretical covariance matrix based of graph G.
}
