% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{comp_joint_mean}
\alias{comp_joint_mean}
\title{Compute the theoretical mean vector based of graph G.}
\usage{
comp_joint_mean(a12, b15, b25, a53, a43, m1, m2, m3, m4, m5)
}
\description{
Compute the theoretical mean vector based of graph G.
}
