% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{comp_pop_ols}
\alias{comp_pop_ols}
\title{Compute the population OLS coefficient (with an intercept)}
\usage{
comp_pop_ols(
  a12,
  b15,
  b25,
  a53,
  a43,
  m1,
  m2,
  m3,
  m4,
  m5,
  s1,
  s2,
  s3,
  s4,
  s5,
  S
)
}
\description{
Compute the population OLS coefficient (with an intercept)
}
