% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algs.R
\name{eliminate_intervals}
\alias{eliminate_intervals}
\title{Remove intervals containing the given indices.}
\usage{
eliminate_intervals(boundary_df, est_ccp)
}
\arguments{
\item{boundary_df}{A data.frame returned by \code{seeded_intervals}.}

\item{est_ccp}{A vector of indices, e.g., estimated causal change points.}
}
\value{
A data.frame with possibly reduced number of rows.
}
\description{
Remove intervals containing the given indices.
}
\details{
This function removes the intervals (rows) in \code{boundary_df} which
contain the indices in \code{est_ccp}.
}
