% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cp_types}
\alias{get_cp_types}
\title{Summarize the types of the change points.}
\usage{
get_cp_types(loc_ccp, loc_nccp)
}
\arguments{
\item{loc_ccp}{A integer vector of time points where causal mechanism changes.}

\item{loc_nccp}{A integer vector of time points where non-causal mechanism
changes.}
}
\value{
A list containing the type(s) of each change point. The types are
"CCP" or "NCCP". The change points are sorted from earliest to latest.
}
\description{
Summarize the types of the change points.
}
\details{
Given a vector of time points of causal and non-causal changes,
sort them and returns the type(s) of each change point. There can be
overlaps between the two sets, which are considered as CCPs by our
definition.
}
