% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{get_full_mb_params}
\alias{get_full_mb_params}
\title{Obtain parameters at each time point.}
\usage{
get_full_mb_params(n, params_step)
}
\arguments{
\item{n}{A scalar of total number of time indices.}

\item{params_step}{A data.frame containing the set of parameters, index
(\code{idx}), and change point types (\code{is_ccp} and \code{is_nccp}).}
}
\description{
Obtain parameters at each time point.
}
