% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{get_one_vs_rest}
\alias{get_one_vs_rest}
\title{Get the list of indices of one-vs-rest after splitting evenly into sub-intervals.}
\usage{
get_one_vs_rest(st, en, n_seg = 3)
}
\arguments{
\item{st}{The start index (inclusive).}

\item{en}{The end index (inclusive).}

\item{n_split}{Number of splits.}
}
\value{
A list of lists where each inner list contains two vectors.
}
\description{
Get the list of indices of one-vs-rest after splitting evenly into sub-intervals.
}
\details{
In the inner list the shorter interval comes first.
}
