% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_powerset}
\alias{get_powerset}
\title{Generate all subsets of a set.}
\usage{
get_powerset(S)
}
\arguments{
\item{S}{A vector with no duplicated values representing a set.}
}
\value{
A list of all subsets of \code{S}.
}
\description{
Generate all subsets of a set.
}
\details{
This function is used to generate all potential invariant sets.
}
