% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{get_rest_vs_one}
\alias{get_rest_vs_one}
\title{Get the list of indices of rest-vs-one after splitting evenly into sub-intervals.}
\usage{
get_rest_vs_one(st, en, n_split = 3)
}
\arguments{
\item{st}{Starting index (inclusive).}

\item{en}{Ending index (inclusive).}

\item{n_split}{}
}
\value{
A list of lists where each inner list contains two vectors.
}
\description{
Get the list of indices of rest-vs-one after splitting evenly into sub-intervals.
}
\details{
In the inner list the longer interval comes first.
}
