% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_seg_lens}
\alias{get_seg_lens}
\title{Calculate the length of each segment.}
\usage{
get_seg_lens(cps_su, n_obs)
}
\arguments{
\item{cps_su}{An integer vector of sorted and unique change points.}

\item{n_obs}{An integer, number of observations.}
}
\value{
An integer vector of \code{length(cps_su) + 1}.
}
\description{
Calculate the length of each segment.
}
