% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_segs}
\alias{get_segs}
\title{Construct segments of indicies based on split points.}
\usage{
get_segs(cps, n_tot)
}
\arguments{
\item{cps}{A vector of integers, the split points (the points themselves are
considered as the starting index of the next segment).}

\item{n_tot}{An integer, total number of indicies.}
}
\value{
A list of length \code{length(cps) + 1}, each containing the indices
in the segment.
}
\description{
Construct segments of indicies based on split points.
}
