% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{get_splits}
\alias{get_splits}
\title{Get the list of indices after splitting an interval evenly into sub-intervals.}
\usage{
get_splits(st, en, n_seg = 3)
}
\arguments{
\item{st}{Starting index (inclusive).}

\item{en}{Ending index (inclusive).}

\item{n_seg}{Number of segments.}
}
\value{
A list of sequence (vector) of indices.
}
\description{
Get the list of indices after splitting an interval evenly into sub-intervals.
}
