% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{init_mb_params}
\alias{init_mb_params}
\title{Initialize the parameter data.frame.}
\usage{
init_mb_params(init_val = 1)
}
\value{
A data.frame containing all initial values of the MB model and three
additional columns: \code{idx}, \code{is_ccp}, and \code{is_nccp}.
}
\description{
Initialize the parameter data.frame.
}
