% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\name{one_vs_rest_test}
\alias{one_vs_rest_test}
\title{Compare multiple environment with one vs rest.}
\usage{
one_vs_rest_test(
  X_seg,
  Y_seg,
  PS,
  rel_grid,
  test_name = "chow1",
  intercept = TRUE,
  rev_one_rest = FALSE,
  two_dir = FALSE,
  save_resids = FALSE
)
}
\arguments{
\item{X_seg}{A (segment of the) covariate matrix.}

\item{Y_seg}{A (segment of the) response vector`.}

\item{PS}{A list of conditioning sets e.g. returned by \code{get_powerset}.}

\item{rel_grid}{A vector of scalars in (0, 1) representing the relative grid, e.g., c(1/3, 2/3)}

\item{test_name}{A string "chow1" or "chow2".}

\item{intercept}{A boolean whether to include an intercept in the regression.}

\item{two_dir}{A boolean whether to test both one vs rest and rest vs one.}

\item{save_resids}{A boolean whether to return residuals.}
}
\value{
A list containing p-value and residuals (which could be NULL if
\code{save_resids} is FALSE).
}
\description{
Compare multiple environment with one vs rest.
}
\details{
Segment the observations according to relative grid point locations and
compare one vs the rest.
}
