% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{rand_mb_params_step}
\alias{rand_mb_params_step}
\title{Generate parameters for the 5 node MB graph.}
\usage{
rand_mb_params_step(
  n,
  true_ccp,
  true_nccp,
  seed_num,
  nccp_mean_nc = 4,
  nccp_mean_sc = 0.5,
  nccp_sig_nc = 4,
  nccp_sig_sc = 0.5,
  nccp_coef_nc = 3,
  nccp_coef_sc = 0.5,
  ccp_mean_nc = 1,
  ccp_mean_sc = 0.5,
  ccp_sig_nc = 1,
  ccp_sig_sc = 0.5,
  ccp_coef_nc = 2,
  ccp_coef_sc = 0.5,
  fixed_sign = 0,
  params_init = NULL
)
}
\arguments{
\item{n}{A scalar of total number of time indices.}

\item{true_ccp}{A vector of integers (time indices) that at least one causal
parameters should change.}

\item{true_nccp}{A vector of integers (time indices) that at least one
non-causal parameters should change.}

\item{seed_num}{An integer as seed number for the random change direction.
Ignored if \code{fixed_sign} is not 0.}

\item{nccp_mean_nc}{Number of non-causal mean params to be changed (up to 4).}

\item{nccp_mean_sc}{The scale of the change applied to each of the params
that should have a change.}

\item{fixed_sign}{A scalar indicating the change direction, can be one of
-1 (only decrease), 1 (only increase), or 0 (random sign). Default to be 0.}
}
\value{
A list of two data.frames. One contains the initial parameters and
one row per change point. Another contains one row per time index
corresponding to the parameters at the time index.
}
\description{
Generate parameters for the 5 node MB graph.
}
\details{
A set of initial values for the causal parameters and non-causal
parameters is first generated based on a given seed number. Then, at each
change point based on its type, a (random sub)set of parameters that
correspond to causal or non-causal change will randomly increase or
decrease a chosen percentage. Note that a \code{true_ccp} refers to time points
where the causal parameters (m5, s5, b15, and b25) change and a \code{true_nccp}
refers to a time point where the non-causal parameters (m1 to m4, s1 to s4,
a12, a53, and a43) change. There can be overlaps between the two and the
overlaps are considered as CCPs by the definition in the paper.
}
