% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algs.R
\name{seedBS_ccp}
\alias{seedBS_ccp}
\title{Seeded binary segmentation for causal change points.}
\usage{
seedBS_ccp(
  X,
  Y,
  boundary_df,
  est_ccp = integer(),
  test_fun = NULL,
  test_param = list(alpha = 0.05),
  esti_fun = NULL,
  esti_param = list(intercept = TRUE),
  verbose = TRUE
)
}
\arguments{
\item{X}{The design matrix of dimension \verb{n_tot x d}.}

\item{Y}{The response vector of length \code{n_tot}.}

\item{test_fun}{A function for testing the existence of a causal change point,
taking values \code{X}, \code{Y}, an interval \code{I}, and possibly other parameters
specified in \code{test_param}, returns a p-value between 0 and 1.}

\item{test_param}{A list containing additional parameters used by \code{test_fun}.}

\item{est_fun}{A loss function for estimating the location of a change point,
taking values \code{X}, \code{Y}, an interval \code{I}, and possibly other parameters
specified in \code{est_param}, returns a change point estimate.}

\item{est_param}{A list containing additional parameters used by \code{loss_fun}.}
}
\value{
A vector of estimated change points.
}
\description{
Seeded binary segmentation for causal change points.
}
\details{
This procedure relies on a test function for whether an interval
contains a change point or not and a estimation function for estimating
the location of a change point.
}
