% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algs.R
\name{seeded_intervals}
\alias{seeded_intervals}
\title{Calculate the boundaries of seeded intervals.}
\usage{
seeded_intervals(n, decay = sqrt(2), min_len = NULL, return_unique = TRUE)
}
\arguments{
\item{n}{Total number of observations.}

\item{decay}{A real number in \verb{(1, 2]} which is the inverse of the decay
parameter in \verb{[1/2, 1)}. Default value is \code{sqrt(2)}.}

\item{min_len}{Minimum length of the intervals. If NULL, all length are
reserved.}

\item{return_unique}{A boolean, whether to return only the unique intervals
or not. Default to be \code{TRUE}.}
}
\value{
A matrix with 2 columns the start and end indices of the seeded
intervals.
}
\description{
Calculate the boundaries of seeded intervals.
}
\references{
Solt Kovács, Housen Li, Peter Buhlmann, and Axel Munk (2022),
Seeded binary segmentation: a general methodology for fast and optimal
changepoint detection, Biometrika, Volume 110, Issue 1, Pages 249–256,
https://doi.org/10.1093/biomet/asac052
}
