% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data from graph G with multivariate Gaussian variables.}
\usage{
sim_data(n, m1, m2, m3, m4, m5, s1, s2, s3, s4, s5, a12, b15, b25, a53, a43)
}
\description{
Simulate data from graph G with multivariate Gaussian variables.
}
\details{
Graph G is a dag where Y (X5) has two parent X1 and X2
and one children X3. X2 is also a child of X1, and X3 is also
a child of X4. This can also be done by using
\code{mvtnorm::rmvnorm(n, mu, Sig)}.
}
