% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_node_example.R
\name{sim_data_mix}
\alias{sim_data_mix}
\title{Simulate data from graph G with mixed multivariate Gaussian variables.}
\usage{
sim_data_mix(
  m1_all,
  m2_all,
  m3_all,
  m4_all,
  m5_all,
  s1_all,
  s2_all,
  s3_all,
  s4_all,
  s5_all,
  a12_all,
  b15_all,
  b25_all,
  a53_all,
  a43_all,
  seed_num
)
}
\description{
Simulate data from graph G with mixed multivariate Gaussian variables.
}
\details{
Slow implementation at the moment.
}
