% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algs.R
\name{stdBS_ccp}
\alias{stdBS_ccp}
\title{Standard binary segmentation for causal change points.}
\usage{
stdBS_ccp(
  X,
  Y,
  st,
  en,
  min_gap = 100,
  test_fun = NULL,
  test_param = list(alpha = 0.05),
  esti_fun = NULL,
  esti_param = list(intercept = TRUE),
  verbose = TRUE
)
}
\value{
A vector of estimated causal change points.
}
\description{
Standard binary segmentation for causal change points.
}
\details{
In the case that the estimation function (\code{esti_fun}) does not
achieve global min at one of the true CCPs when multiple CCPs exist, the
returned vector may contain time points that are not CCPs. One may use
the \code{prune_candi} function to remove the NCCPs.
}
