% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chow_tests.R
\name{test_chow1}
\alias{test_chow1}
\title{Chow test (eq 29 in Chow 1960)}
\usage{
test_chow1(X, Y, I1, I2, S, intercept = TRUE)
}
\value{
A list of the test statistic and the p-value.
}
\description{
Chow test (eq 29 in Chow 1960)
}
